--- a/libraries/lib-time-and-pitch/StaffPad/VectorOps.h
+++ b/libraries/lib-time-and-pitch/StaffPad/VectorOps.h
@@ -14,6 +14,15 @@
 #include <cstdint>
 #include <cstring>
 
+#if defined(_MSC_VER) && (defined(_M_AMD64) || defined(_M_X64) || \
+                          (defined(_M_IX86_FP) && _M_IX86_FP >= 2))
+#define USE_SSE2_COMPLEX 1
+#endif
+
+#if USE_SSE2_COMPLEX
+#   include "SimdComplexConversions_sse2.h"
+#endif
+
 namespace staffpad {
 namespace vo {
 
@@ -90,6 +99,32 @@ inline void findMaxElement(const T* src, int32_t n, int32_t& maxIndex, T& maxVal
   }
 }
 
+#if USE_SSE2_COMPLEX
+
+inline void calcPhases(const std::complex<float>* src, float* dst, int32_t n)
+{
+  simd_complex_conversions::perform_parallel_simd_aligned(
+     src, dst, n,
+     [](const __m128 rp, const __m128 ip, __m128& out)
+     { out = simd_complex_conversions::atan2_ps(ip, rp); });
+}
+
+inline void calcNorms(const std::complex<float>* src, float* dst, int32_t n)
+{
+  simd_complex_conversions::perform_parallel_simd_aligned(
+     src, dst, n,
+     [](const __m128 rp, const __m128 ip, __m128& out)
+     { out = simd_complex_conversions::norm(rp, ip); });
+}
+
+inline void rotate(
+   const float* oldPhase, const float* newPhase, std::complex<float>* dst,
+   int32_t n)
+{
+  simd_complex_conversions::rotate_parallel_simd_aligned(
+     oldPhase, newPhase, dst, n);
+}
+#else
 inline void calcPhases(const std::complex<float>* src, float* dst, int32_t n)
 {
   for (int32_t i = 0; i < n; i++)
@@ -109,6 +144,7 @@ inline void rotate(const float* oldPhase, const float* newPhase, std::complex<fl
     dst[i] *= std::complex<float>(cosf(theta), sinf(theta));
   }
 }
+#endif
 
 } // namespace vo
 } // namespace staffpad