This is an attempt to improve the performance of the time stretcher on MSVC. The problem was noticed due to extreme duration required to run the time stretching test.

MSVC, especially when targeting 32 bit Windows has the following problems:

* `rintf` is not inlined. This function is approximately 2 CPU instructions. Replacing it with `int(x + 0.5f)` results in 2x performance improvement.
* Math functions are not inlined either. `void TimeAndPitch::_process_hop` first converts a large array of complex numbers (2049 were observed during the tests) to polar coordinates, manipulates the result and converts back to cartesian coordinates. This conversion turned out to be extremely slow on MSVC.

On top of that, Audacity build system only enabled SSE2 support for `SBSMS` and `soundtouch`.

As the result of this PR, the following improvement was observed on Ryzen 5800x:

* Win64: 938 seconds -> 215 seconds
* Win32: 5489 (!) seconds -> 354 seconds

This result is quite close to M1-powered MacBook Pro, where the test runs in around 200 seconds.

Checking the samples used for unit tests showed that the difference is below -90 dB.

### For QA

This PR changes how trigonometry functions are calculated: `sin`, `cos` and `atan`. The difference between the "reference" implementation and implementation is relatively small, however, the error accumulates. I have observed a difference of around -90 dB, I expect that something similar will be observed during the testing.

------------------

- [x] I signed [CLA](https://www.audacityteam.org/cla/)
- [x] The title of the pull request describes an issue it addresses
- [x] If changes are extensive, then there is a sequence of easily reviewable commits
- [x] Each commit's message describes its purpose and effects
- [x] There are no behavior changes unnecessary for the stated purpose of the PR

Recommended:
- [x] Each commit compiles and runs on my machine without known undesirable changes of behavior