@saintmatthieu 2023-10-04
I'd give this PR the thumbs up, but I'm incompetent to review the SSE optimization part, and I understood that this was going to be expected from @vsverchinsky.