@crsib 2023-10-04
Thank you, @vsverchinsky! Your suggestions are now implemented