@chinakov 2023-10-05
Tested and discussed with @crsib - we've come to the conclusion that the performance is satisfactory.